+++ Include/Acidanthera/Library/OcAppleKernelLib.h
@@ -125,6 +125,7 @@ typedef enum KERNEL_CACHE_TYPE_ {
 #define KERNEL_VERSION_MOJAVE_MIN           KERNEL_VERSION (18, 0, 0)
 #define KERNEL_VERSION_CATALINA_MIN         KERNEL_VERSION (19, 0, 0)
 #define KERNEL_VERSION_BIG_SUR_MIN          KERNEL_VERSION (20, 0, 0)
+#define KERNEL_VERSION_MONTEREY_MIN         KERNEL_VERSION (21, 0, 0)
 
 //
 // Maximum kernel versions for each release.
@@ -141,6 +142,7 @@ typedef enum KERNEL_CACHE_TYPE_ {
 #define KERNEL_VERSION_HIGH_SIERRA_MAX      (KERNEL_VERSION_MOJAVE_MIN - 1)
 #define KERNEL_VERSION_MOJAVE_MAX           (KERNEL_VERSION_CATALINA_MIN - 1)
 #define KERNEL_VERSION_CATALINA_MAX         (KERNEL_VERSION_BIG_SUR_MIN - 1)
+#define KERNEL_VERSION_BIG_SUR_MAX          (KERNEL_VERSION_MONTEREY_MIN - 1)
 
 //
 // Prelinked context used for kernel modification.