+++ Changelog.md
@@ -7,6 +7,7 @@ OpenCore Changelog
 - Fixed `PowerTimeoutKernelPanic` on macOS 12
 - Fixed transparency click detection on OpenCanopy boot entries
 - Added PCI device info dumping to `SysReport`
+- Fixed `SetApfsTrimTimeout` on macOS 12
 
 #### v0.7.0
 - Fixed NVRAM reset on firmware with write-protected `BootOptionSupport`